requirements/requirements.txt
pdfminer3k==1.3.1
jieba==0.42.1
matplotlib==3.3.2
wordcloud==1.8.1
numpy==1.19.2
pdfminer==20191125
Pillow==8.1.0
PySimpleGUI==4.34.0
